pyproject.toml
[tool.poetry]
name = "django_oauthapi"
version = "0.1.0"
description = ""
authors = ["defalt91 <armansoltanian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-oauth-toolkit = "^1.5.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
